+++ src/components/molecules/CheckboxesList/__test__/NewCheckboxesList.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import NewCheckboxesList from '../NewCheckboxesList';
+import NewCheckboxesList from '../CheckboxesList';
 
 it('has a titled group', () => {
   // ARRANGE
