+++ src/components/molecules/CheckboxesList/index.ts
@@ -1,0 +1,1 @@
+export { default as CheckboxesList } from './CheckboxesList';
