--- src/components/molecules/NewCheckboxesList/index.ts
@@ -1,1 +1,0 @@
-export { default as NewCheckboxesList } from './NewCheckboxesList';
